Refactor: Split api.go into multiple smaller files
